+++ Modules/Tasks/Database/Migrations/2024_11_03_141510_create_user_task_rewards_table.php
@@ -11,6 +11,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('user_task_rewards')) {
+            return;
+        }
         Schema::create('user_task_rewards', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
